DocumentTranslation/docconvertorfunction-pdf2image/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pdf2image==1.16.3
img2pdf==0.4.4
azure-storage-blob==12.6.0
